packages/slack/requirements.txt
paho-mqtt>=1.5.0
schema>=0.6.8
slackclient==1.3.1